+++ NEWS.md
@@ -1,3 +1,7 @@
+# fracture (development version)
+
+* Add `is.frac_mat()`.
+
 # fracture 0.1.0
 
 * Initial release.